refactor: move auth server start logic out of medusa.c
