+++ security/medusa/include/l4/auth_server.h
@@ -1,7 +1,8 @@
 #ifndef _MEDUSA_AUTH_SERVER_H
 #define _MEDUSA_AUTH_SERVER_H
 
-void wait_for_auth_server(void);
 void set_auth_server_ready(void);
+void wait_for_auth_server(void);
+void start_auth_server(void);
 
 #endif
